eduvoice_backend/requirements.txt
# Django Core
Django==4.2.7
django-environ==0.11.2
python-dotenv==1.0.0
setuptools>=65.5.1

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
django-cors-headers==4.3.0

# API Documentation
drf-spectacular==0.26.5

# Database
psycopg2-binary==2.9.9

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.5.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Text-to-Speech
gTTS==2.4.0
elevenlabs==0.2.27
pydub==0.25.1
mutagen==1.47.0

# Email
django-anymail==10.2

# Security
django-ratelimit==4.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
